+++ src/components/box/box.config.ts
@@ -1,3 +1,5 @@
+import { space } from "../../style/themes/base/base-theme.config";
+
 export default {
   light: {
     thumb: "var(--colorsUtilityMajor300)",
@@ -7,5 +9,13 @@ export default {
   dark: {
     thumb: "var(--colorsUtilityMajor200)",
     track: "var(--colorsUtilityMajor400)",
+  },
+
+  gap: (gapValue: number | string) => {
+    if (typeof gapValue === "number") {
+      return space[gapValue];
+    }
+
+    return gapValue;
   },
 };